+++ packages/node/src/integrations/anr/index.ts
@@ -1,8 +1,10 @@
+import { types } from 'node:util';
 import { Worker } from 'node:worker_threads';
 import type { Contexts, Event, EventHint, Integration, IntegrationFn, ScopeData } from '@sentry/core';
 import {
   GLOBAL_OBJ,
   defineIntegration,
+  getClient,
   getCurrentScope,
   getFilenameToDebugIdMap,
   getGlobalScope,
@@ -12,7 +14,10 @@ import {
 } from '@sentry/core';
 import { NODE_VERSION } from '../../nodeVersion';
 import type { NodeClient } from '../../sdk/client';
+import { isDebuggerEnabled } from '../../utils/debug';
 import type { AnrIntegrationOptions, WorkerStartData } from './common';
+
+const { isPromise } = types;
 
 // This string is a placeholder that gets overwritten with the worker code.
 export const base64WorkerScript = '###AnrWorkerScript###';
@@ -94,8 +99,13 @@ const _anrIntegration = ((options: Partial<AnrIntegrationOptions> = {}) => {
         });
       }
     },
-    setup(initClient: NodeClient) {
+    async setup(initClient: NodeClient) {
       client = initClient;
+
+      if (options.captureStackTrace && (await isDebuggerEnabled())) {
+        logger.warn('ANR captureStackTrace has been disabled because the debugger was already enabled');
+        options.captureStackTrace = false;
+      }
 
       // setImmediate is used to ensure that all other integrations have had their setup called first.
       // This allows us to call into all integrations to fetch the full context
@@ -212,4 +222,27 @@ async function _startWorker(
     worker.terminate();
     clearInterval(timer);
   };
+}
+
+export function disableAnrDetectionForCallback<T>(callback: () => T): T;
+export function disableAnrDetectionForCallback<T>(callback: () => Promise<T>): Promise<T>;
+/**
+ * Disables ANR detection for the duration of the callback
+ */
+export function disableAnrDetectionForCallback<T>(callback: () => T | Promise<T>): T | Promise<T> {
+  const integration = getClient()?.getIntegrationByName(INTEGRATION_NAME) as AnrInternal | undefined;
+
+  if (!integration) {
+    return callback();
+  }
+
+  integration.stopWorker();
+
+  const result = callback();
+  if (isPromise(result)) {
+    return result.finally(() => integration.startWorker());
+  }
+
+  integration.startWorker();
+  return result;
 }