+++ packages/node/test/integration/transactions.test.ts
@@ -620,7 +620,7 @@ describe('Integration | Transactions', () => {
     expect(logs).toEqual(
       expect.arrayContaining([
         'SpanExporter dropped 2 spans because they were pending for more than 300 seconds.',
-        'SpanExporter exported 1 spans, 0 unsent spans remaining',
+        'SpanExporter exported 1 spans, 0 spans are waiting for their parent spans to finish',
       ]),
     );
   });
@@ -680,7 +680,7 @@ describe('Integration | Transactions', () => {
 
     expect(spans).toHaveLength(2);
 
-    expect(spans[0]!.description).toEqual('inner span 1');
-    expect(spans[1]!.description).toEqual('inner span 2');
+    expect(spans).toContainEqual(expect.objectContaining({ description: 'inner span 1' }));
+    expect(spans).toContainEqual(expect.objectContaining({ description: 'inner span 2' }));
   });
 });